// src/app/models/house.model.ts
export class House {
    draw17: boolean = false;
    holecard: boolean = false;
    doubleonsoft: boolean = true;
    splitacesone: boolean = true;
    splitacesnobjs: boolean = true;
    resplitaces: boolean = true;
    surrender: boolean = false;
    surrendervsace: boolean = false;
    surrenderearly: boolean = false;
    decks: number = 6;
    maxsplits: number = 0;
}